ws_data/requirements.txt
Django==3.1.5
psycopg2-binary>=2.8
djangorestframework==3.12.2
django-filter
requests
pyyaml
django-rest-swagger



